Zsh: `_themis.vim` isn’t a completion file but zinit is confused

+++ .config/zsh/utility-functions.zsh
@@ -711,18 +711,18 @@ function zsh:plugins:update {
 
   # Use subshell to restore pwd.
   (
-    # Clean up the directory because enhancd makes it dirty when loaded.
-    execute_with_echo builtin cd "${ENHANCD_ROOT:?}"
-    execute_with_echo git restore .
+    # Clean up the directory before updating.
+    execute_with_echo builtin cd "${ENHANCD_ROOT:?}" && execute_with_echo git restore .
+    execute_with_echo builtin cd "${THEMIS_HOME:?}"  && execute_with_echo git restore .
   )
 
   execute_with_echo zinit update --all --parallel --quiet
 
   # Use subshell to restore pwd.
   (
-    # Remove `_*.fish` files because they are treated as completions by zinit.
-    execute_with_echo builtin cd "${ENHANCD_ROOT:?}"
-    execute_with_echo rm -f ./**/_*.fish
+    # Remove `_*` files because they are treated as completions by zinit.
+    execute_with_echo builtin cd "${ENHANCD_ROOT:?}" && execute_with_echo rm -f ./**/_*.fish
+    execute_with_echo builtin cd "${THEMIS_HOME:?}"  && execute_with_echo rm -f ./**/_*.vim
   )
 
   execute_with_echo zinit creinstall "${ZINIT[BIN_DIR]}"
